[P4ADEV-1700] fixed merge
