+++ openapi/generated.openapi.json
@@ -10,8 +10,8 @@
     "description" : "Generated server url"
   } ],
   "tags" : [ {
-    "name" : "DebtPositionType",
-    "description" : "the DebtPositionType API"
+    "name" : "DebtPositionTypes",
+    "description" : "the DebtPositionTypes API"
   }, {
     "name" : "Taxonomy",
     "description" : "the Taxonomy API"
@@ -509,9 +509,80 @@
         } ]
       }
     },
-    "/bff/debt-position-types/{organizationId}" : {
+    "/bff/debt-position-types/{id}" : {
       "get" : {
-        "tags" : [ "DebtPositionType" ],
+        "tags" : [ "DebtPositionTypes" ],
+        "summary" : "Retrieve a specific debt position type",
+        "operationId" : "getDebtPositionType",
+        "parameters" : [ {
+          "name" : "id",
+          "in" : "path",
+          "description" : "The unique identifier of the debt position type",
+          "required" : true,
+          "schema" : {
+            "type" : "string"
+          }
+        } ],
+        "responses" : {
+          "200" : {
+            "description" : "OK",
+            "content" : {
+              "application/json" : {
+                "schema" : {
+                  "$ref" : "#/components/schemas/DebtPositionTypeDTO"
+                }
+              }
+            }
+          },
+          "404" : {
+            "description" : "Not Found - Debt position type not found",
+            "content" : {
+              "application/json" : {
+                "schema" : {
+                  "$ref" : "#/components/schemas/DebtPositionTypeDTO"
+                }
+              }
+            }
+          },
+          "401" : {
+            "description" : "Unauthorized - Missing or invalid token",
+            "content" : {
+              "application/json" : {
+                "schema" : {
+                  "$ref" : "#/components/schemas/DebtPositionTypeDTO"
+                }
+              }
+            }
+          },
+          "403" : {
+            "description" : "Forbidden - Access denied",
+            "content" : {
+              "application/json" : {
+                "schema" : {
+                  "$ref" : "#/components/schemas/DebtPositionTypeDTO"
+                }
+              }
+            }
+          },
+          "500" : {
+            "description" : "Internal Server Error",
+            "content" : {
+              "application/json" : {
+                "schema" : {
+                  "$ref" : "#/components/schemas/DebtPositionTypeDTO"
+                }
+              }
+            }
+          }
+        },
+        "security" : [ {
+          "bearerAuth" : [ ]
+        } ]
+      }
+    },
+    "/bff/debt-position-types/organization/{organizationId}" : {
+      "get" : {
+        "tags" : [ "DebtPositionTypes" ],
         "summary" : "Retrieve a page of DebtPositionTypeWithCount based on the organization's brokerId",
         "operationId" : "getDebtPositionTypeWithCount",
         "parameters" : [ {
@@ -875,6 +946,66 @@
           "description" : {
             "type" : "string",
             "description" : "Description of the error"
+          }
+        }
+      },
+      "DebtPositionTypeDTO" : {
+        "required" : [ "brokerId", "code", "collectingReason", "debtPositionTypeId", "description", "flagAnonymousFiscalCode", "flagMandatoryDueDate", "flagNotifyIo", "macroArea", "orgType", "serviceType", "taxonomyCode" ],
+        "type" : "object",
+        "properties" : {
+          "debtPositionTypeId" : {
+            "type" : "integer",
+            "description" : "Unique identifier for the debt position type",
+            "format" : "int64"
+          },
+          "brokerId" : {
+            "type" : "integer",
+            "description" : "Identifier for the broker associated with this debt position type",
+            "format" : "int64"
+          },
+          "code" : {
+            "type" : "string",
+            "description" : "Code representing the debt position type"
+          },
+          "description" : {
+            "type" : "string",
+            "description" : "Description of the debt position type"
+          },
+          "orgType" : {
+            "type" : "string",
+            "description" : "Organization type associated with this debt position type"
+          },
+          "macroArea" : {
+            "type" : "string",
+            "description" : "Macro area related to the debt position type"
+          },
+          "serviceType" : {
+            "type" : "string",
+            "description" : "Service type associated with the debt position type"
+          },
+          "collectingReason" : {
+            "type" : "string",
+            "description" : "Reason for collecting this debt position"
+          },
+          "taxonomyCode" : {
+            "type" : "string",
+            "description" : "Taxonomy code for this debt position type"
+          },
+          "flagAnonymousFiscalCode" : {
+            "type" : "boolean",
+            "description" : "Indicates if an anonymous fiscal code is allowed"
+          },
+          "flagMandatoryDueDate" : {
+            "type" : "boolean",
+            "description" : "Indicates if a due date is mandatory"
+          },
+          "flagNotifyIo" : {
+            "type" : "boolean",
+            "description" : "Indicates if notifications to IO are required"
+          },
+          "ioTemplateMessage" : {
+            "type" : "string",
+            "description" : "Template message for IO notifications"
           }
         }
       },
