+++ src/main/java/it/gov/pagopa/pu/bff/connector/debt_position/config/DebtPositionApisHolder.java
@@ -3,6 +3,7 @@ package it.gov.pagopa.pu.bff.connector.debt_position.config;
 import it.gov.pagopa.pu.debtpositions.controller.ApiClient;
 import it.gov.pagopa.pu.debtpositions.controller.BaseApi;
 import it.gov.pagopa.pu.debtpositions.controller.generated.DebtPositionTypeEntityControllerApi;
+import it.gov.pagopa.pu.debtpositions.controller.generated.DebtPositionTypeWithCountSearchControllerApi;
 import jakarta.annotation.PreDestroy;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
@@ -15,6 +16,7 @@ import org.springframework.web.client.RestTemplate;
 public class DebtPositionApisHolder {
 
   private final DebtPositionTypeEntityControllerApi debtPositionTypeEntityControllerApi;
+  private final DebtPositionTypeWithCountSearchControllerApi debtPositionTypeWithCountSearchControllerApi;
   private final ThreadLocal<String> bearerTokenHolder = new ThreadLocal<>();
 
   public DebtPositionApisHolder(
@@ -26,6 +28,7 @@ public class DebtPositionApisHolder {
     apiClient.setBearerToken(bearerTokenHolder::get);
 
     this.debtPositionTypeEntityControllerApi = new DebtPositionTypeEntityControllerApi(apiClient);
+    this.debtPositionTypeWithCountSearchControllerApi = new DebtPositionTypeWithCountSearchControllerApi(apiClient);
   }
 
   @PreDestroy
@@ -38,6 +41,11 @@ public class DebtPositionApisHolder {
    */
   public DebtPositionTypeEntityControllerApi getDebtPositionTypeControllerApi(String accessToken) {
     return getApi(accessToken, debtPositionTypeEntityControllerApi);
+  }
+
+  /** It will return a {@link DebtPositionTypeWithCountSearchControllerApi} instrumented with the provided accessToken. Use null if auth is not required */
+  public DebtPositionTypeWithCountSearchControllerApi getDebtPositionTypeWithCountSearchControllerApi(String accessToken){
+    return getApi(accessToken, debtPositionTypeWithCountSearchControllerApi);
   }
 
   private <T extends BaseApi> T getApi(String accessToken, T api) {
