--- src/main/java/it/gov/pagopa/pu/bff/service/debtposition/DebtPositionTypeService.java
@@ -1,12 +1,0 @@
-package it.gov.pagopa.pu.bff.service.debtposition;
-
-import it.gov.pagopa.pu.bff.dto.generated.PagedDebtPositionTypeWithCount;
-import it.gov.pagopa.pu.p4paauth.dto.generated.UserInfo;
-import org.springframework.data.domain.Pageable;
-
-public interface DebtPositionTypeService {
-
-  PagedDebtPositionTypeWithCount getDebtPositionTypeWithCount(
-    Long organizationId, Pageable pageable,
-    UserInfo loggedUser, String accessToken);
-}
