--- src/test/java/it/gov/pagopa/pu/bff/connector/debtposition/config/DebtPositionApisHolderTest.java
@@ -1,50 +1,0 @@
-package it.gov.pagopa.pu.bff.connector.debtposition.config;
-
-import it.gov.pagopa.pu.bff.connector.BaseApiHolderTest;
-import it.gov.pagopa.pu.debtpositions.controller.ApiClient;
-import it.gov.pagopa.pu.debtpositions.dto.generated.PagedModelDebtPositionTypeWithCount;
-import java.util.Collections;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.web.util.DefaultUriBuilderFactory;
-
-@ExtendWith(MockitoExtension.class)
-class DebtPositionApisHolderTest extends BaseApiHolderTest {
-  @Mock
-  private RestTemplateBuilder restTemplateBuilderMock;
-
-  private DebtPositionApisHolder debtPositionApisHolder;
-
-  @BeforeEach
-  void setUp() {
-    Mockito.when(restTemplateBuilderMock.build()).thenReturn(restTemplateMock);
-    Mockito.when(restTemplateMock.getUriTemplateHandler()).thenReturn(new DefaultUriBuilderFactory());
-    ApiClient apiClient = new ApiClient(restTemplateMock);
-    String baseUrl = "http://example.com";
-    apiClient.setBasePath(baseUrl);
-    debtPositionApisHolder = new DebtPositionApisHolder(baseUrl, restTemplateBuilderMock);
-  }
-
-  @AfterEach
-  void verifyNoMoreInteractions() {
-    Mockito.verifyNoMoreInteractions(
-      restTemplateBuilderMock,
-      restTemplateMock
-    );
-  }
-
-  @Test
-  void whenGetDebtPositionTypeWithCountSearchControllerApiThenAuthenticationShouldBeSetInThreadSafeMode() throws InterruptedException {
-    assertAuthenticationShouldBeSetInThreadSafeMode(
-      accessToken -> debtPositionApisHolder.getDebtPositionTypeWithCountSearchControllerApi(accessToken)
-        .crudDebtPositionTypesWithCountFindByBrokerId(1L,0,0, Collections.emptyList()),
-      PagedModelDebtPositionTypeWithCount.class,
-      debtPositionApisHolder::unload);
-  }
-}
