+++ src/test/java/it/gov/pagopa/pu/bff/controller/DebtPositionTypeControllerTest.java
@@ -1,12 +1,5 @@
 package it.gov.pagopa.pu.bff.controller;
 
-import it.gov.pagopa.pu.bff.dto.generated.DebtPositionTypeWithCount;
-import it.gov.pagopa.pu.bff.dto.generated.PagedDebtPositionTypeWithCount;
-import java.time.OffsetDateTime;
-import java.util.List;
-import org.junit.jupiter.api.Assertions;
-import it.gov.pagopa.pu.bff.dto.generated.DebtPositionTypeDTO;
-import it.gov.pagopa.pu.bff.service.debt_position.DebtPositionTypeService;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -15,8 +8,10 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import it.gov.pagopa.pu.bff.dto.generated.DebtPositionTypeDTO;
 import it.gov.pagopa.pu.bff.dto.generated.DebtPositionTypeWithCount;
 import it.gov.pagopa.pu.bff.dto.generated.PagedDebtPositionTypeWithCount;
+import it.gov.pagopa.pu.bff.service.debt_position.DebtPositionTypeService;
 import java.time.OffsetDateTime;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
@@ -26,9 +21,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -36,11 +29,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class DebtPositionTypeControllerTest {
@@ -50,6 +38,8 @@ class DebtPositionTypeControllerTest {
 
   @InjectMocks
   private DebtPositionTypeController debtPositionTypeController;
+
+  private DebtPositionTypeDTO debtPositionTypeDTO;
 
   @BeforeEach
   void setUp() {
