+++ src/test/java/it/gov/pagopa/pu/bff/service/DebtPositionTypeServiceImplTest.java
@@ -1,20 +1,32 @@
 package it.gov.pagopa.pu.bff.service;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertSame;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+
 import it.gov.pagopa.pu.bff.connector.debt_position.client.DebtPositionTypeClient;
 import it.gov.pagopa.pu.bff.dto.generated.DebtPositionTypeDTO;
+import it.gov.pagopa.pu.bff.dto.generated.PagedDebtPositionTypeWithCount;
 import it.gov.pagopa.pu.bff.mapper.DebtPositionTypeDTOMapper;
+import it.gov.pagopa.pu.bff.mapper.DebtPositionTypeWithCountMapper;
 import it.gov.pagopa.pu.bff.service.debt_position.DebtPositionTypeServiceImpl;
 import it.gov.pagopa.pu.debtpositions.dto.generated.DebtPositionType;
+import it.gov.pagopa.pu.debtpositions.dto.generated.PagedModelDebtPositionTypeWithCount;
+import it.gov.pagopa.pu.p4paauth.dto.generated.UserInfo;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.security.authorization.AuthorizationDeniedException;
 
 @ExtendWith(MockitoExtension.class)
 class DebtPositionTypeServiceImplTest {
@@ -24,6 +36,10 @@ class DebtPositionTypeServiceImplTest {
 
   @Mock
   private DebtPositionTypeDTOMapper debtPositionTypeDTOMapperMock;
+  @Mock
+  private AuthorizationService authorizationServiceMock;
+  @Mock
+  private DebtPositionTypeWithCountMapper debtPositionTypeWithCountMapperMock;
 
   private DebtPositionTypeServiceImpl debtPositionTypeService;
 
@@ -65,7 +81,7 @@ class DebtPositionTypeServiceImplTest {
       .ioTemplateMessage("Test IO Template Message")
       .build();
 
-    debtPositionTypeService = new DebtPositionTypeServiceImpl(debtPositionTypeClientMock, debtPositionTypeDTOMapperMock);
+    debtPositionTypeService = new DebtPositionTypeServiceImpl(debtPositionTypeClientMock, debtPositionTypeDTOMapperMock, debtPositionTypeWithCountMapperMock, authorizationServiceMock);
   }
 
   @Test
@@ -103,5 +119,44 @@ class DebtPositionTypeServiceImplTest {
     assertNull(result);
   }
 
+
+  @Test
+  void givenValidUserWhenGetDebtPositionTypeWithCountThenOK() {
+    long brokerId = 1L;
+    UserInfo userInfo = new UserInfo();
+    userInfo.setBrokerId(brokerId);
+    PagedModelDebtPositionTypeWithCount pagedModelDebtPositionTypeWithCount = new PagedModelDebtPositionTypeWithCount();
+    PagedDebtPositionTypeWithCount pagedDebtPositionTypeWithCount = new PagedDebtPositionTypeWithCount();
+
+    Mockito.doNothing().when(authorizationServiceMock).validateAdminRole(1L,userInfo);
+    Mockito.when(debtPositionTypeClientMock.getDebtPositionTypeWithCount(brokerId, PageRequest.of(0,10),accessToken)).thenReturn(pagedModelDebtPositionTypeWithCount);
+    Mockito.when(debtPositionTypeWithCountMapperMock.mapToPagedDebtPositionWithCount(pagedModelDebtPositionTypeWithCount)).thenReturn(pagedDebtPositionTypeWithCount);
+
+    PagedDebtPositionTypeWithCount result = debtPositionTypeService.getDebtPositionTypeWithCount(
+      1L, PageRequest.of(0,10),
+      userInfo, accessToken);
+
+    assertNotNull(result);
+    assertSame(pagedDebtPositionTypeWithCount,result);
+
+    Mockito.verifyNoMoreInteractions(debtPositionTypeClientMock,debtPositionTypeWithCountMapperMock,authorizationServiceMock);
+  }
+
+  @Test
+  void givenInvalidUserWhenGetDebtPositionTypeWithCountThenAuthorizationDeniedException() {
+    long brokerId = 1L;
+    UserInfo userInfo = new UserInfo();
+    userInfo.setBrokerId(brokerId);
+    PageRequest pageRequest = PageRequest.of(0, 10);
+
+    Mockito.doThrow(new AuthorizationDeniedException("")).when(authorizationServiceMock).validateAdminRole(1L,userInfo);
+
+    Assertions.assertThrows(AuthorizationDeniedException.class,()->
+      debtPositionTypeService.getDebtPositionTypeWithCount(
+        1L, pageRequest, userInfo, accessToken));
+
+    Mockito.verifyNoMoreInteractions(authorizationServiceMock);
+    Mockito.verifyNoInteractions(debtPositionTypeClientMock,debtPositionTypeWithCountMapperMock);
+  }
 }
